WIP: ComponentController/Presentation REFACTORED
