--- src/main/java/ecdar/utility/helpers/MouseTrackable.java
@@ -1,7 +1,0 @@
-package ecdar.utility.helpers;
-
-import ecdar.utility.mouse.MouseTracker;
-
-public interface MouseTrackable extends LocationAware {
-    MouseTracker getMouseTracker();
-}
